requirements.txt
Flask==3.1.0
flask-cors==5.0.1
numpy==2.1.2
onnx==1.17.0
onnxruntime==1.20.1
pillow==11.0.0